pyproject.toml
[tool.poetry]
name = "micro-cinema"
version = "0.1.0"
description = ""
authors = ["yilbegan <w2v@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Pyrogram = "^2.0.35"
py-tgcalls = "^0.9.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
dataclass-factory = "^2.16"
pydantic = "^1.9.1"
TgCrypto = "^1.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
